Move image schema outside of campaign
