+++ src/schema/image/focal-point.js
